🎉 Completed Delete and Update functionality

+++ Kurirska Služba/Views/Prices.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Kurirska_Služba.Forms;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
@@ -26,6 +27,7 @@ namespace Kurirska_Služba.Views
             InitializeComponent();
             ShowData();
         }
+
         public void ShowData()
         {
             string sqlSelect = @"select CenaID as 'Cena ID', Opis, Tezina, Cena from tblCenovnik";
@@ -33,11 +35,51 @@ namespace Kurirska_Služba.Views
             dgPrices.ItemsSource = dataTable.DefaultView;
         }
 
+        private void btnChange_Click(object sender, RoutedEventArgs e)
+        {
+            if (dgPrices.SelectedItem != null)
+            {
+                DataRowView selectedRow = (DataRowView)dgPrices.SelectedItems[0];
+                WindowPrice window = new WindowPrice(Convert.ToInt32(selectedRow["Cena ID"])) { Owner = Application.Current.MainWindow };
+                window.ShowDialog();
+                ShowData();
+            }
+            else
+            {
+                MessageBox.Show("Morate selektovati vrednost iz liste kako bi je menjali", "Element nije izabran", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+        }
+
         private void btnDelete_Click(object sender, RoutedEventArgs e)
         {
-            DataRowView id = (DataRowView)dgPrices.SelectedItems[0];
-            DatabaseConnection.DeleteById(id["Cena ID"].ToString(),"CenaID", "tblCenovnik");
-            ShowData();
+            if (dgPrices.SelectedItem != null)
+            {
+                DataRowView id = (DataRowView)dgPrices.SelectedItems[0];
+                DatabaseConnection.DeleteById(id["Cena ID"].ToString(), "CenaID", "tblCenovnik");
+                ShowData();
+            }
+            else
+            {
+                MessageBox.Show("Morate selektovati vrednost iz liste kako bi je obrisali", "Element nije izabran", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+        }
+
+        private void dgPrices_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (dgPrices.SelectedItem != null)
+            {
+                foreach (Button button in spControls.Children)
+                {
+                    button.IsEnabled = true;
+                }
+            }
+            else
+            {
+                foreach (Button button in spControls.Children)
+                {
+                    button.IsEnabled = true;
+                }
+            }
         }
     }
 }
